rework BackpressureObjectCounter
